Add opcache preloading support
