+++ .gitignore
@@ -3,3 +3,4 @@
 /tmp
 /vendor
 /tests/Fixtures/var/
+/.idea
