+++ tests/Functional/FunctionalTest.php
@@ -67,7 +67,7 @@ class FunctionalTest extends WebTestCase
         $this->assertSame('<a href="http://localhost/employee/95">FirstName94 LastName94</a>', $sample->link);
 
         // Change when we drop old PHP versions and thus old PHPunit versions
-        $this->assertRegExp('#href="/employee/[0-9]+"#', $sample->buttons);
+        $this->assertMatchesRegularExpression('#href="/employee/[0-9]+"#', $sample->buttons);
         //$this->assertMatchesRegularExpression('#href="/employee/[0-9]+"#', $sample->buttons);
         $this->assertSame('04-07-2016', $json->data[6]->employedSince);
     }
